% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermal.R
\name{anderson}
\alias{anderson}
\title{Calculates thermal time based on the Anderson model}
\usage{
anderson(Tvec, Tb, To, Tc, sum = TRUE)
}
\arguments{
\item{Tvec}{A vector of hourly temperatures, in degrees C.}

\item{Tb}{The base cardinal temperature of the model.}

\item{To}{The optimal cardinal temperature of the model.}

\item{Tc}{The critical cardinal temperature of the model.}

\item{sum}{Logical, should the vector of thermal times be returned as a
sum?}
}
\value{
either the number or vector of numbers representing the thermal
    time calculated
}
\description{
This function calculates thermal time in growing degree hours from a
    vector of temperatures (in degrees C) based on the Anderson
    \emph{et al.} (1986) model of thermal time described below.
}
\details{
The functional form of the Anderson thermal time model is as
    follows:
    \deqn{ GDH = \begin{cases}
        0 & T\leq T_b \\
        \frac{T_o-T_a}{2} \left[1+\cos\left(\pi + \pi \cdot \frac{T-T_b}{T_o-T_b}\right) \right] & T_b\leq T\leq T_o \\
        (T_o-T_a) \left[1+\cos\left(\frac{\pi}{2} + \frac{\pi}{2} \cdot \frac{T-T_o}{T_c-T_o}\right) \right] & T_o\leq T \leq T_c \\
        0 & T_c \leq T
        \end{cases}}
    where \eqn{T} is the hourly temperature, \eqn{T_b} is the base
    temperature, \eqn{T_o} is the optimal temperature, and \eqn{T_c} is
    the critical temperature.

     Anderson, J. L., E. A. Richardson, and C. D. Kesner. 1985.
         "Validation of Chill Unit and Flower Bud Phenology Models
         for 'Montmorency' sour Cherry." In I International Symposium on
         Computer Modelling in Fruit Research and Orchard Management 184,
         71-78. http://www.actahort.org/books/184/184_7.htm.
}
\examples{
temp <- seq(-5, 50)
gdh <- anderson(temp, 4, 25, 36, sum=FALSE)
plot(gdh ~ temp)
}

