% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterfunctions.R
\name{boundlength}
\alias{boundlength}
\title{Length of the bounds variables}
\usage{
boundlength(form, CT, L, lim)
}
\arguments{
\item{form}{character, the thermal time functional form.}

\item{CT}{logical, do the cardinal temperatures need to be estimated?}

\item{L}{logical, does the thermal time accumulation length need to be
estimated?}

\item{lim}{logical list, do the start/ends need to be estimated?}
}
\value{
The length the bounds vectors need to be.
}
\description{
This function returns the length that the lbounds and ubounds variables
    should be based on the functinoal form of the thermal time
    accumulation as well as whether or not the cardinal temperatures and
    the length of accumulation need to be estimated.
}
