% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterfunctions.R
\name{calclength}
\alias{calclength}
\title{Calculates length of model}
\usage{
calclength(ml, lims)
}
\arguments{
\item{ml}{numeric, the lengths of each of the stages, or NA}

\item{lims}{list, the start/stop pairs for each of the stages.}
}
\value{
A vector with the lengths of each of the stages.
}
\description{
Calculates length of model
}
