% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R
\docType{methods}
\name{cardinaltemps<-}
\alias{cardinaltemps<-}
\alias{cardinaltemps<-,ParameterList-method}
\title{Setting the cardinal temps}
\usage{
cardinaltemps(object) <- value

\S4method{cardinaltemps}{ParameterList}(object) <- value
}
\arguments{
\item{object}{An object of class ParameterList}

\item{value}{A list of cardinal temperatures}
}
\description{
Used to change the cardinal temperature parameters without recreating the
      object.
}

