% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R,
%   R/plantmodelmethods.R, R/flowermodelmethods.R
\docType{methods}
\name{cardinaltemps}
\alias{cardinaltemps}
\alias{cardinaltemps,ParameterList-method}
\alias{cardinaltemps,PlantModel-method}
\alias{cardinaltemps,FlowerModel-method}
\title{Returns the returns a list of cardinal temperatures}
\usage{
cardinaltemps(object)

\S4method{cardinaltemps}{ParameterList}(object)

\S4method{cardinaltemps}{PlantModel}(object)

\S4method{cardinaltemps}{FlowerModel}(object)
}
\arguments{
\item{object}{An object of class ParameterList}
}
\value{
A list of cardinal temperatures, one for each stage of the model.
}
\description{
Returns the returns a list of cardinal temperatures

Accesses the cardinaltemps list of a ParameterList object

Accesses the cardinal temperatures

Accesses the cardinal temperatures
}
