% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterfunctions.R
\name{checkpars}
\alias{checkpars}
\title{Checks to see if parameters are in the correct order}
\usage{
checkpars(pars)
}
\arguments{
\item{pars}{a vector of cardinal temperatures, should be organized least to greatest}
}
\value{
logical, TRUE if the cardinal temperatures are organized least to greatest, FALSE otherwise
}
\description{
Checks to see if parameters are in the correct order
}

