% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelcheck.R
\name{checktemps}
\alias{checktemps}
\title{Checks temp data.frame is ship shape}
\usage{
checktemps(temp, pheno, forms, modelclass)
}
\arguments{
\item{temp}{data.frame, contains all the temperature data}

\item{pheno}{data.frame, contains all the phenological data}

\item{forms}{list, contains all the functional forms to be used to calculate
the thermal time in the model.}

\item{modelclass}{character, specifies the class of the model being fit,
either 'PlantModel' or 'FlowerModel'}
}
\value{
logical, TRUE if everything goes well, an error if not.
}
\description{
Checks temp data.frame is ship shape
}
