% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{checktempyears}
\alias{checktempyears}
\title{Checks year presence in Plant object}
\usage{
checktempyears(object)
}
\arguments{
\item{object}{An object of the class Plant}
}
\value{
Logical. Returns TRUE if all years of phenology data also have
    corresponding temperature data. Otherwise it returns the years that are
    missing temperature data.
}
\description{
This function checks to make sure that for every year where there is
    phenology data there is also temperature data.
}

