% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermal.R
\name{chillbasic}
\alias{chillbasic}
\title{Calculates winter chill based on the basic chill unit}
\usage{
chillbasic(tvec, Tmax, sum = TRUE)
}
\arguments{
\item{tvec}{numeric, a vector of temperatures used to calculate chill}

\item{sum}{logical, should the accumulated chill be summed?}
}
\value{
A numeric vector of chill values or a single chill value if sum is
    TRUE.
}
\description{
This function calculates the amount of chill accumulated based on the 32F-45F (0C-7.2C) model.
}
\details{
This model is described in Luedeling, E. and Brown, P.H., 2011. A
    global analysis of the comparability of winter chill models for fruit
    and nut trees. International Journal of Biometeorology, 55(3),
    pp.411-421.
}
