% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fithelper.R
\name{convertParameters}
\alias{convertParameters}
\title{Separates out Parameter Values for optimization}
\usage{
convertParameters(pars, modtype, S, TH, vp, eventvec, years, modclass)
}
\arguments{
\item{pars}{numeric, a vector of parameters that are optimized using the
`DEoptim` function.}

\item{modtype}{character, is the model a day threshold ('DT') model or a
thermal time threshold model ('TTT').}

\item{S}{logical/numeric, is the day to start counting being optimized? If
not, the value associated with the start day should be provided.}

\item{TH}{logical/numeric, should the model threshold be optimized. If
not L is the (numeric) model threshold.}

\item{vp}{character, c('start', 'threshold') should either of these pars
vary from year to year.}

\item{eventvec}{numeric, a vector that contains the day of the year for the
starting event of the model. For bloom the event is harvest.}
}
\value{
list of the correct start POSIXct date vector and the threshold
    vector based on the type of model being run. If the day of the threshold
    varies by year it is a 'Period' vector. If it does not, it is a POSIXct
    date vector.
}
\description{
Assigns parameter values for start and threshold based on model type,
    which parameters are being estimated and which parameters are varying
    from year to year.
}
