% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fithelper.R
\name{convertToMatrix}
\alias{convertToMatrix}
\title{Converts vectors to matrices}
\usage{
convertToMatrix(x, nstage, nform)
}
\arguments{
\item{x}{numeric, object to be converted to a matrix}

\item{nstage}{numeric, number of stages in the model}

\item{nform}{numeric, number of forms in the model}
}
\value{
a matrix with the appropriate number of rows and columns
}
\description{
This function converts vectors to matrices with the appropriate number of
    rows or columns depending on the number of stages and functional forms
    in a model.
}
