% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fithelper.R
\name{correctYears}
\alias{correctYears}
\title{Identifies which year each day corresponds to}
\usage{
correctYears(day, years, modclass)
}
\arguments{
\item{day}{numeric, a vector of start or threshold days}

\item{years}{numeric, a vector of years of the phenological event
associated with the start days}

\item{modclass}{character, PlantModel or FlowerModel}
}
\value{
A vector of years that correspond to the actual year each of the days
    in the day vector happened.
}
\description{
This function decides whether a day happened in the same year as the
    associated phenological event, or the year before.
}
