% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{crossval}
\alias{crossval}
\title{Crossvalidates the PlantModel Error}
\usage{
crossval(plant)
}
\arguments{
\item{plant}{PlantModel, the object to be crossvalidated}

\item{k}{numeric, number of folds for crossvalidation.}
}
\value{
A PlantModel object with the error slot having the new crossvalidated
    error
}
\description{
Does k-fold crossvalidation on the error of a PlantModel object.
}

