% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantmodelclass.R, R/plantmodelmethods.R
\docType{methods}
\name{crossvalidated<-}
\alias{crossvalidated<-}
\alias{crossvalidated<-,ParameterList-method}
\title{Setting whether the error is crossvalidated}
\usage{
crossvalidated(object) <- value

\S4method{crossvalidated}{ParameterList}(object) <- value
}
\arguments{
\item{object}{An object of class PlantModel}

\item{value}{logical, is the error crossvalidated?}
}
\description{
Used to change whether or not the error is crossvalidated without recreating
      the object.
}

