% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantmodelclass.R, R/plantmodelmethods.R,
%   R/flowermodelmethods.R
\docType{methods}
\name{crossvalidated}
\alias{crossvalidated}
\alias{crossvalidated,PlantModel-method}
\alias{crossvalidated,FlowerModel-method}
\title{Returns the crossvalidated logical vecter}
\usage{
crossvalidated(object)

\S4method{crossvalidated}{PlantModel}(object)

\S4method{crossvalidated}{FlowerModel}(object)
}
\arguments{
\item{object}{An object of class PlantModel}
}
\value{
A logical vector answering the question 'Is the model error
    crossvalidated?'
}
\description{
Returns the crossvalidated logical vecter

Is the model error crossvalidated?

Is the model error crossvalidated?
}
