% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{crossvalTEST}
\alias{crossvalTEST}
\title{Crossvalidates the PlantModel Error}
\usage{
crossvalTEST(flower, temps, k, seed, fun = "rmse", lbounds, ubounds,
  iterations = 100, cores = 1L)
}
\arguments{
\item{flower}{flowermodel, the object to be crossvalidated}

\item{temps}{data.frame, contains all the temperature data.}

\item{k}{numeric, number of folds for crossvalidation.}

\item{seed}{numeric, the number to be used as the seed. Set to NA for no seed
set.}

\item{fun}{character, the name of the function used to evaluate the model.}

\item{lbounds}{numeric, lower bounds of parameters}

\item{ubounds}{numeric, upper bounds of parameters.}

\item{iterations}{numeric, number of iterations to be used when optimizing
paramters.}

\item{cores}{integer, number of cores to be used when fitting the models}
}
\value{
A PlantModel object with the error slot having the new crossvalidated
    error
}
\description{
Does k-fold crossvalidation on the error of a PlantModel object.
}
