% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{daylength}
\alias{daylength}
\title{Calculates the length of a day}
\usage{
daylength(jul, lat = 38.5, ret = "")
}
\arguments{
\item{jul}{The julian day/day of the year (1-366)}

\item{lat}{The latitude of the location}

\item{ret}{I am unsure what this argument is for}
}
\value{
The length of the day in hours, along with the time of sunrise and
    sunset.
}
\description{
This function calculates the length of a day given the day of the year and
    the latitude of the location.
}
