% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fithelper.R
\name{dayToDAE}
\alias{dayToDAE}
\title{Converts day of the year to day after event (DAE)}
\usage{
dayToDAE(day, event, years)
}
\arguments{
\item{day}{numeric, the day of the year you want to convert to days after
event}

\item{event}{numeric, the day of the event for each year of interest}

\item{years}{numeric, the year that each event happened}
}
\value{
A numeric vector specifying the number of days after the event that
   the day of interest happened.
}
\description{
This function takes in a day of the year and the day of a phenological event
    and converts the day of the year to the number of days after the
    phenological event. The year of each event must be specified so that the
    appropriate year length can be used.
}
