% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fithelper.R
\name{dayToDate}
\alias{dayToDate}
\title{Converts day of the year to date-time}
\usage{
dayToDate(years, days, modclass, varying, hours = NA,
  startTime = "00:00:00", timezone = "America/Los_Angeles")
}
\arguments{
\item{years}{numeric, a vector of years to use for each day of the year}

\item{days}{numeric, the day of the year you want to convert to date-time.}

\item{modclass}{character, 'PlantModel' or 'FlowerModel'.}

\item{varying}{character,}

\item{hours}{numeric, a vector of hours to use in your date-time conversion}

\item{startTime}{character, used to specify the same time of day for each
date-time entry element. Only used if hours=NA.}
}
\value{
A POSIXct vector of date-times
}
\description{
This function takes in a vector of days of the year and a vector of years
    and converts them to a date-time vector of the class POSIXct
}
