% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantmodelclass.R, R/plantmodelmethods.R
\docType{methods}
\name{error<-}
\alias{error<-}
\alias{error<-,PlantModel-method}
\title{Setting the error}
\usage{
error(object) <- value

\S4method{error}{PlantModel}(object) <- value
}
\arguments{
\item{object}{An object of class PlantModel}

\item{value}{numeric, the model error}
}
\description{
Used to change the error without recreating the object.
}

