% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantmodelclass.R, R/plantmodelmethods.R
\docType{methods}
\name{error}
\alias{error}
\alias{error,PlantModel-method}
\title{Returns the RMSE for a phenological model}
\usage{
error(object)

\S4method{error}{PlantModel}(object)
}
\arguments{
\item{object}{An object of class PlantModel}
}
\value{
The root mean squared error (RMSE) of the model
}
\description{
Returns the RMSE for a phenological model

Accesses the RMSE of a PlantModel object
}
