% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R
\docType{methods}
\name{estimate<-}
\alias{estimate<-}
\alias{estimate<-,ParameterList-method}
\title{Setting the parameters to estimate}
\usage{
estimate(object) <- value

\S4method{estimate}{ParameterList}(object) <- value
}
\arguments{
\item{object}{An object of class ParameterList}

\item{value}{A character vector containing either 'cardinaltemps',
'modlength' or both.}
}
\description{
Used to change the functional form without recreating the object.
}

