% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/plantmodelmethods.R
\docType{methods}
\name{estimate}
\alias{estimate}
\alias{estimate,PlantModel-method}
\title{Returns the parameters to estimate}
\usage{
estimate(object)

\S4method{estimate}{PlantModel}(object)
}
\arguments{
\item{object}{An object of class ParameterList}
}
\value{
character, the names of the parameters to be estimated
    (cardinaltemps, modlength or both).
}
\description{
Returns the parameters to estimate

Accesses what parameters are to be estimated
}

