% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{extractParameters}
\alias{extractParameters}
\title{Extracts Parameters from optimization}
\usage{
extractParameters(estimate, parname, parlist, optlist)
}
\arguments{
\item{estimate}{list, list of the following logical vectors: estimatestart,
estimatethresh, estimateCT}

\item{parname}{character, the name of the parameter to extract, either
'start', 'threshold', or 'cardinaltemps'}

\item{parlist}{list, list of parameterlists}

\item{optlist}{list, list of optim function outputs}
}
\value{
the value of the selected parameter
}
\description{
This function extracts either the start day, threshold or cardinal
    temperatures from the parameterlist or optim object depending on whether
    the parameter was optimized/estimated or not.
}
