% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractResults.R
\name{extractResults}
\alias{extractResults}
\title{Gets phenologial data and season length predictions}
\usage{
extractResults(pm, cultivars, form)
}
\description{
Gets phenologial data and season length predictions
}
