% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowermodel.R
\name{flowermodel}
\alias{flowermodel}
\title{Fits phenological models}
\usage{
flowermodel(phenology, temps, parlist, lbounds, ubounds, cores = 1L,
  iterations = 200)
}
\arguments{
\item{phenology}{data.frame, contains the phenological data.
Required columns are `year`, `event1` and `event0` which represents the
harvest date from the previous year.}

\item{temps}{data.frame, contains the temperature data. There should be a
column for year, day of the year, hourly temperature (named temp) or min
max daily temperatures (named tmin and tmax), and hour if the
temperatures are hourly. Must have data from one year before you have
phenology data}

\item{parlist}{ParameterList, contains the parameter values and
functional form of the thermal time calculations.}

\item{lbounds}{numeric, a vector of lower bounds for the parameters in the
model}

\item{ubounds}{numeric, a vector of the upper bounds for the parameters in
the model.}

\item{cores}{integer, if using parallel processing, how many cores should R
use to fit the model.}

\item{iterations}{numeric, the number of iterations used in the differential
evolution optimization of the phenological parameters.}
}
\value{
A FlowerModel object.
}
\description{
`flowermodel` fits phenological models of two general types: thermal time accumulation and day accumulation (for more information see ____).
}
