% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/plantmethodsbasic.R
\docType{methods}
\name{form<-}
\alias{form<-}
\alias{form<-,Plant-method}
\title{Setting the form}
\usage{
form(object) <- value

\S4method{form}{Plant}(object) <- value
}
\arguments{
\item{object}{A Plant object}

\item{value}{A function that represents the functional form
of the thermal time model.}
}
\description{
Used to change the thermal time functional form without recreating the
     object.
}

