% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fithelper.R
\name{formatParameters}
\alias{formatParameters}
\title{Formats parameters in days to Dates}
\usage{
formatParameters(years, eventday, startday, threshold, modtype, modclass,
  varying)
}
\arguments{
\item{years}{numeric, a vector of years to use for each day of the year.}

\item{eventday}{numeric, a vector that contains the day of the year for the
starting event of the model. For bloom the event is harvest.}

\item{startday}{numeric, vector of days to start count thermal time each
year.}

\item{threshold}{numeric, the threshold of the model.}

\item{modtype}{character, is the model a day threshold ('DT') model or a
thermal time threshold model ('TTT').}

\item{modclass}{character, PlantModel or FlowerModel}

\item{varying}{character, c('start', 'threshold') should either of these pars
vary from year to year.}
}
\value{
list of the correct start POSIXct date vector and the threshold
    vector based on the type of model being run. If the day of the threshold
    varies by year it is a 'Period' vector. If it does not, it is a POSIXct
    date vector.
}
\description{
Assigns parameter values for start and threshold based on model type,
    which parameters are being estimated and which parameters are varying
    from year to year.
}
