% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R
\docType{methods}
\name{forward<-}
\alias{forward<-}
\alias{forward<-,ParameterList-method}
\title{Setting whether the model counts forward from bloom}
\usage{
forward(object) <- value

\S4method{forward}{ParameterList}(object) <- value
}
\arguments{
\item{object}{An object of class ParameterList}

\item{value}{logical, does the model count forward from bloom?}
}
\description{
Used to change whether the model counts forward from bloom (TRUE) or
     backward from bloom (FALSE).
}
