% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R
\name{forward}
\alias{forward}
\title{Returns whether the model is forward or backward}
\usage{
forward(object)
}
\arguments{
\item{object}{An object of class ParameterList}
}
\value{
logical, whether the model counts forward from bloom or not.
}
\description{
Returns whether the model is forward or backward
}
