% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermal.R
\name{gdd}
\alias{gdd}
\title{Calculates thermal time based on the growing degree day (GDD) model}
\usage{
gdd(tdat, Tb, sum = TRUE)
}
\arguments{
\item{tdat}{A matrix of minimum and maximum daily temperatures, in degrees
C.}

\item{Tb}{The base cardinal temperature of the model.}

\item{sum}{Logical, should the vector of thermal times be returned as a
sum?}
}
\value{
either the number or vector of numbers representing the thermal
    time calculated
}
\description{
This function calculates thermal time in growing degree days from a
    vector of temperatures (in degrees C) based on the model
    described below.
}
\details{
The functional form of the flat thermal time model is as follows:
    \deqn{GDD = \begin{cases}
        0 & T_{max}\leq T_b \\
        \frac{(T_{max}-T_b)^2}{2(T_{max} - T_{min})} & T_{min} \leq T_b \leq T_{max}\\
        T - T_{avg} & T_{min} \geq T_b\\
        \end{cases}}
    where \eqn{T_min} is the daily minimum temperature, \eqn{T_max} is the
    daily maximum temperature, and \eqn{T_b} is the base temperature.
}
\examples{
temp <- data.frame(tmin=runif(10, 0, 10), tmax=runif(10, 13, 25))
gdd(temp, 4, sum=TRUE)
gdd(temp, 4, sum=FALSE)
}

