% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermal.R
\name{gddsimple}
\alias{gddsimple}
\title{Calculates thermal time based on the simplified growing degree day (GDD)
     model}
\usage{
gddsimple(tdat, Tb, sum = TRUE)
}
\arguments{
\item{tdat}{A matrix of minimum and maximum daily temperatures, in degrees
C.}

\item{Tb}{The base cardinal temperature of the model.}

\item{sum}{Logical, should the vector of thermal times be returned as a
sum?}
}
\value{
either the number or vector of numbers representing the thermal
    time calculated
}
\description{
This function calculates thermal time in growing degree days from a
    vector of temperatures (in degrees C) based on the model
    described below.
}
\details{
The functional form of the flat thermal time model is as follows:
    \deqn{GDD = \begin{cases}
        0 & T_{avg}\leq T_b \\
        T - T_{avg} & T_b \leq T_{avg}\\
        \end{cases}}
    where \eqn{T_{avg}} is the average daily temperature, and \eqn{T_b} is
    the base temperature.
}
\examples{
temp <- data.frame(tmin=runif(10, 0, 10), tmax=runif(10, 13, 25))
gddsimple(temp, 4, sum=TRUE)
gddsimple(temp, 4, sum=FALSE)
}

