% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{head.list}
\alias{head.list}
\title{Return the First of an Object}
\usage{
\method{head}{list}(x, n = 6L, ...)
}
\arguments{
\item{x}{a list that you want to see the first n entries of}

\item{n}{integer, number of entries of the list you want to see}

\item{...}{anything else you can past to head}
}
\description{
Return the First of an Object
}
