% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{is.leapyear}
\alias{is.leapyear}
\title{Checks if a year is a leap year}
\usage{
is.leapyear(year)
}
\arguments{
\item{year}{a numeric vector of years to test}
}
\value{
A logical vector. True indicates the year is a leap year.
}
\description{
Checks if a year is a leap year
}
\examples{
y <- seq(1990, 2010)
is.leapyear(y)
}
