% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{isDateTime}
\alias{isDateTime}
\title{Is X a date or time object?}
\usage{
isDateTime(x)
}
\description{
This function tests if the object x is a date or time object (Period or
    POSIXct).
}
\details{
@param x vector, the object you want to test
 @value logical, is the object one of the date time classes specified above.
 @export
}
