% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowerPlant.R
\name{lastDay}
\alias{lastDay}
\title{Returns the end day of a stage}
\usage{
lastDay(ModelClass, stage)
}
\arguments{
\item{ModelClass}{character, the type of model used, either PlantModel or
FlowerModel}

\item{stage}{numeric, what stage is the is the model predicting for.}
}
\value{
The name of the response variable as a character.
}
\description{
This function takes in the type of model (PlantModel or FlowerModel) and
    the stage and based on that decides the name of the column of the last
    day of the stage.
}
