% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{length0}
\alias{length0}
\title{Checks if length of vector is zero}
\usage{
length0(x)
}
\arguments{
\item{x}{vector or list}
}
\value{
logical, TRUE if the vector is of length 0, FALSE if it has a length
    greater than zero
}
\description{
Checks if length of vector is zero
}

