% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R
\docType{methods}
\name{limits<-}
\alias{limits<-}
\alias{limits<-,ParameterList-method}
\title{Setting the modlength}
\usage{
limits(object) <- value

\S4method{limits}{ParameterList}(object) <- value
}
\arguments{
\item{object}{An object of class ParameterList}

\item{value}{A list of start/stop pairs}
}
\description{
Used to change the limits without recreating the object.
}
