% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R
\docType{methods}
\name{limits}
\alias{limits}
\alias{limits,ParameterList-method}
\title{Returns the returns a vector of start/stop pairs}
\usage{
limits(object)

\S4method{limits}{ParameterList}(object)
}
\arguments{
\item{object}{An object of class ParameterList}
}
\value{
A list of the start/stop pairs of the model
}
\description{
Returns the returns a vector of start/stop pairs

Accesses the model limits of the ParameterList object
}
