% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermal.R
\name{linear}
\alias{linear}
\title{Calculates thermal time based on the linear model}
\usage{
linear(Tvec, Tb, sum = TRUE)
}
\arguments{
\item{Tvec}{A vector of hourly temperatures, in degrees C.}

\item{Tb}{The base cardinal temperature of the model.}

\item{sum}{Logical, should the vector of thermal times be returned as a
sum?}
}
\value{
either the number or vector of numbers representing the thermal
    time calculated
}
\description{
This function calculates thermal time in growing degree hours from a
    vector of temperatures (in degrees C) based on the linear model
    described below.
}
\details{
The functional form of the flat thermal time model is as follows:
    \deqn{GDH = \begin{cases}
    0 & T\leq T_b \\
    T - T_b & T_b \leq T
    \end{cases}}
    where \eqn{T} is the hourly temperature, \eqn{T_b} is the base
    temperature and \eqn{T_o} is the optimal temperature.
}
\examples{
temp <- seq(-5, 50)
gdh <- linear(temp, 4, sum=FALSE)
plot(gdh ~ temp)
}

