% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R
\name{mclass<-}
\alias{mclass<-}
\title{Setting stages as PlantModels or a FlowerModels}
\usage{
mclass(object) <- value
}
\arguments{
\item{object}{An object of class ParameterList}

\item{value}{logical, does the model count forward from bloom?}
}
\description{
Used to change whether the model uses the PlantModel class or the
 FlowerModel class.
}
