% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R
\name{mclass}
\alias{mclass}
\title{Returns whether the model is a 'PlantModel' or a 'FlowerModel'}
\usage{
mclass(object)
}
\arguments{
\item{object}{An object of class ParameterList}
}
\value{
character, whether the model is a PlantModel or a FlowerModel
}
\description{
Returns whether the model is a 'PlantModel' or a 'FlowerModel'
}
