% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minrmse.R
\name{minrmse}
\alias{minrmse}
\title{Calculates model RMSE}
\usage{
minrmse(pars, fdat, tdat, modtype, form, stage, CT, L)
}
\arguments{
\item{pars}{Cardinal temperatures}

\item{fdat}{the data.frame containing the phenological information}

\item{tdat}{list containing the temperature information}

\item{modtype}{character, the type of model. Can be 'partial', 'full' or
'combined'.}

\item{form}{the functional form of the thermal time accumulation}

\item{stage}{the number of the stage of the phenological model}

\item{CT}{logical, should the cardinal temperatures be optimized. If not CT
is the vector of cardinal temperatures}

\item{L}{logical, should the model length be optimized. If not L is the
model length.}

\item{length}{the length of thermal time accumulation (in days). It can be
either a set length of time (one number) or the total length of the
stage (one length for each entry in fdat).}
}
\value{
The RMSE value for a given set of cardinal temperatures and thermal
    time accumulation length.
}
\description{
Calculates RMSE for a partial, full or combined model of thermal time
    accumulation given a certain set of cardinal temperatures and thermal
    time accumulation length.
}

