% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minrmse.R
\name{minrmseTTT}
\alias{minrmseTTT}
\title{Calculates TTT model RMSE}
\usage{
minrmseTTT(pars, fdat, tdat, form, length, stage)
}
\arguments{
\item{pars}{Cardinal temperatures}

\item{fdat}{the data.frame containing the phenological information}

\item{tdat}{list containing the temperature information}

\item{form}{the functional form of the thermal time accumulation}

\item{length}{the length of thermal time accumulation (in days). It can be
either a set length of time (one number) or the total length of the
stage (one length for each entry in fdat).}

\item{stage}{the number of the stage of the phenological model}
}
\value{
The RMSE value for a given set of cardinal temperatures and thermal
    time accumulation length.
}
\description{
Calculates RMSE for the combined model of thermal time accumulation given a
    certain set of cardinal temperatures and thermal time accumulation
    length.
}

