% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantclass.R, R/plantmethodsbasic.R
\docType{methods}
\name{modeltype<-}
\alias{modeltype<-}
\alias{modeltype<-,Plant-method}
\title{Setting the model type}
\usage{
modeltype(object) <- value

\S4method{modeltype}{Plant}(object) <- value
}
\arguments{
\item{object}{A Plant object}

\item{value}{A character that specifies the type of model. Current options
options are partial, full, combined, and time.}
}
\description{
Used to change the model type without recreating the object.
}

