% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantmodelcheck.R
\name{modeltypecheck}
\alias{modeltypecheck}
\title{Checks model type}
\usage{
modeltypecheck(mt)
}
\arguments{
\item{mt}{character, the type of model to be fit using plantmodel()}
}
\value{
The function returns true if the model type is one of the allowed
    types (thermal or day), and false with a message if it is not.
}
\description{
Checks model type
}

