% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/parameterlistmethods.R
\docType{methods}
\name{modlength<-}
\alias{modlength<-}
\alias{modlength<-,ParameterList-method}
\title{Setting the modlength}
\usage{
modlength(object, ...) <- value

\S4method{modlength}{ParameterList}(object) <- value
}
\arguments{
\item{object}{An object of class ParameterList}
}
\value{
A vector of model lengths
}
\description{
Used to change the model length without recreating the object.
}

