% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R,
%   R/plantmodelmethods.R, R/flowermodelmethods.R
\docType{methods}
\name{modlength}
\alias{modlength}
\alias{modlength,ParameterList-method}
\alias{modlength,PlantModel-method}
\alias{modlength,FlowerModel-method}
\title{Returns the returns a vector of accumulation lengths}
\usage{
modlength(object)

\S4method{modlength}{ParameterList}(object)

\S4method{modlength}{PlantModel}(object)

\S4method{modlength}{FlowerModel}(object)
}
\arguments{
\item{object}{An object of class ParameterList}
}
\value{
A list of model lengths
}
\description{
Returns the returns a vector of accumulation lengths

Accesses the modlength vector of a ParameterList object

Accesses the model lengths of a plant object

Accesses the model lengths of a FlowerModel object
}
