% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectivefunction.R
\name{objective}
\alias{objective}
\title{Generates the Objective function}
\usage{
objective(parlist, phenology, templist, stage, CT, Start, Threshold, listindex,
  mclass)
}
\arguments{
\item{parlist}{ParameterList, contains all of the parlist needed to
create the phenology model.}

\item{phenology}{data.frame, contains all of the phenology data necessary for
running the phenology model.}

\item{templist}{list, contains all the temperature data for estimating
thermal time.}

\item{stage}{numeric, the stage of the model.}

\item{CT}{Should the cardinal temperatures be optimized? If yes CT is TRUE,
if not CT should be a vector of the cardinal temperatures.}

\item{Start}{logical, Should model start day be optimized? If yes,
Start is TRUE, if not Start should be the start day to be
used.}

\item{Threshold}{logical, Should model threshold be optimized? If yes,
Threshold is TRUE, if not Threshold should be the model threshold to be
used.}

\item{listindex}{numeric, the index of the ParameterList from parlist that
you are at}

\item{mclass}{character, type of model to be estimating, options are
'PlantModel' or 'FlowerModel'. If you have negative day values, you
probably want flower model.}

\item{simple}{logical, is the simplified version of the model being run?}
}
\value{
the function that is passed to DEoptim to optimize.
}
\description{
This function creates a function to be passed to the DEoptim function that
optimizes the cardinal temperatures and the length of thermal time or day
accumulation for a PlantModel or FlowerModel object.
}
