% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectivefunction.R
\name{objectiveFlower}
\alias{objectiveFlower}
\title{Generates the Objective function for a Flower Model}
\usage{
objectiveFlower(parlist, phenology, templist, stage, CT, Threshold, simple,
  listindex, startday)
}
\arguments{
\item{parlist}{ParameterList, contains all of the parlist needed to
create the phenology model/PlantModel object.}

\item{phenology}{data.frame, contains all of the phenology data necessary for
running the phenology model.}

\item{templist}{list, contains all the temperature data for estimating
thermal time.}

\item{stage}{numeric, the stage of the model.}

\item{CT}{Should the cardinal temperatures be optimized? If yes CT is TRUE,
if not CT should be a vector of the cardinal temperatures.}

\item{Threshold}{logical, Should model threshold be optimized? If yes,
Threshold is TRUE, if not Threshold should be the model threshold to be
used.}

\item{simple}{logical, is the simplified version of the model being run?}

\item{listindex}{numeric, the index of the ParameterList from parlist that
you are at}

\item{startday}{logical, is the day to start counting being optimized? (ie.
not using bloom as startday)}
}
\value{
the function that is passed to DEoptim to optimize.
}
\description{
This function creates a function to be passed to the DEoptim function that
optimizes the cardinal temperatures and the length of thermal time or day
accumulation for a FlowerModel object.
}
