% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectivefunction.R
\name{objectivePlant}
\alias{objectivePlant}
\title{Generates the Objective function for a PlantModel}
\usage{
objectivePlant(parlist, phenology, templist, stage, CT, Start, Threshold,
  listindex)
}
\arguments{
\item{parlist}{ParameterList, contains all of the parlist needed to
create the phenology model/PlantModel object.}

\item{phenology}{data.frame, contains all of the phenology data necessary for
running the phenology model.}

\item{templist}{list, contains all the temperature data for estimating
thermal time.}

\item{stage}{numeric, the stage of the model.}

\item{CT}{Should the cardinal temperatures be optimized? If yes CT is TRUE,
if not CT should be a vector of the cardinal temperatures.}

\item{Start}{logical/numeric, is the day to start counting being optimized?
If it is, put TRUE, if not Start should be the starting day to be used.}

\item{Threshold}{logical, Should model threshold be optimized? If yes,
Threshold is TRUE, if not Threshold should be the model threshold to be
used.}

\item{listindex}{numeric, the index of the ParameterList from parlist that
you are at}
}
\value{
the function that is passed to DEoptim to optimize.
}
\description{
This function creates a function to be passed to the DEoptim function that
optimizes the cardinal temperatures and the length of thermal time or day
accumulation for a PlantModel object.
}
