% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{ParameterList-class}
\alias{ParameterList-class}
\title{An S4 class to store model parameters}
\description{
ParameterList stores parameters for running the phenological models
}
\section{Slots}{

\describe{
\item{\code{modlength}}{Stores the length of time or thermal time that is accumulated
in the model. There should be either one value (if there is only one
or if all the stages have the same parameters) or as many values as there
are stages.}

\item{\code{cardinaltemps}}{The cardinal temperatures for the model. This is a list
that should either be of length one or the length of the number of
stages. Each element of the list should contain the same number of
cardinal parameters.}
}}

