% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{parameterlist}
\alias{parameterlist}
\title{Creates ParameterList object}
\usage{
parameterlist(n, mt, simple, ff, ct, modelthreshold, start, varyingparameters,
  optimized = c("cardinaltemps", "threshold", "start"),
  ModelClass = "PlantModel")
}
\arguments{
\item{n}{integer, the number of stages in the model.}

\item{mt}{character, the model type. Can be either 'DT' or 'TTT'.}

\item{simple}{logical, is the model the simplified version?}

\item{ff}{character, the functional form of the thermal time calculations.
Options are 'gdd', 'gddsimple', 'linear', 'flat', 'triangle', and
'anderson'. For more information see ______.}

\item{ct}{A list, data.frame, or matrix of cardinal temperatures. In the
case of a data.frame or matrix, each row should contain the cardinal
temperatures for a given stage of the model. If ct is a list, each
element of the list should be a vector of cardinal temperatures for a
given stage of the model.}

\item{modelthreshold}{numeric vector of model thresholds (one for each
stage).}

\item{start}{numeric vector of start days (one for each stage).}

\item{varyingparameters}{character vector, which parameters will have the
of the year vary from year to year. Can contain,'start' and 'threshold',
or be NA.}

\item{optimized}{character, Determines what parameters are optimized in the
model. Can contain "cardinaltemps", "threshold", "start" or any
combination of all three, but it must contain at least one.}

\item{ModelClass}{character, the class to be fit. Options are
'PlantModel' or 'FlowerModel'.}
}
\value{
An object of the class ParameterList.
}
\description{
This function creates an object of the class ParameterList. Plant objects
    require objects of the class ParameterList for their parameters slot.
}
