% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/plantmethodsbasic.R
\docType{methods}
\name{parameters<-}
\alias{parameters<-}
\alias{parameters<-,Plant-method}
\title{Setting the initial thermal time/day accumulation length}
\usage{
parameters(object) <- value

\S4method{parameters}{Plant}(object) <- value
}
\arguments{
\item{object}{A Plant object}

\item{value}{The length of thermal time accumulation in the model.}
}
\description{
Used to change the number of days or GDD/GDH the model accumulates thermal
     time in the phenological model without recreating the object.
}

