% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantmodelclass.R, R/plantmodelmethods.R,
%   R/flowermodelmethods.R
\docType{methods}
\name{parameters}
\alias{parameters}
\alias{parameters,PlantModel-method}
\alias{parameters,FlowerModel-method}
\title{Returns the parameters for a phenological model}
\usage{
parameters(object)

\S4method{parameters}{PlantModel}(object)

\S4method{parameters}{FlowerModel}(object)
}
\arguments{
\item{object}{An object of class PlantModel}
}
\value{
The ParameterList of the phenological model
}
\description{
Returns the parameters for a phenological model

Accesses the parameters of a PlantModel object

Accesses the parameters of a FlowerModel object
}
