% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterfunctions.R
\name{parnum}
\alias{parnum}
\title{Functional form parameter number}
\usage{
parnum(form)
}
\arguments{
\item{form}{character, the name of the functional form in question.
Options are 'gdd', 'gddsimple', 'linear', 'flat', 'asymcur, 'anderson'
and 'triangle'.}
}
\value{
The number of parameters required for the thermal time functional
    form supplied.
}
\description{
This function returns the number of parameters a supplied thermal time
    functional form (form) requires.
}

