% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R
\docType{methods}
\name{parsOptimized<-}
\alias{parsOptimized<-}
\alias{parsOptimized<-,ParameterList-method}
\title{Setting the parameters to optimize}
\usage{
parsOptimized(object) <- value

\S4method{parsOptimized}{ParameterList}(object) <- value
}
\arguments{
\item{object}{An object of class ParameterList}

\item{value}{A character vector containing either 'cardinaltemps',
'modlength' or both.}
}
\description{
Used to change the functional form without recreating the object.
}

