% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R, R/plantmodelmethods.R
\docType{methods}
\name{parsOptimized}
\alias{parsOptimized}
\alias{parsOptimized,ParameterList-method}
\alias{parsOptimized,PlantModel-method}
\title{Returns the parameters to estimate}
\usage{
parsOptimized(object)

\S4method{parsOptimized}{ParameterList}(object)

\S4method{parsOptimized}{PlantModel}(object)
}
\arguments{
\item{object}{An object of class ParameterList}
}
\value{
character, the names of the parameters to be estimated
    (cardinaltemps, modlength or both).
}
\description{
Returns the parameters to estimate

Accesses which parameters to estimate for the ParameterList object

Accesses what parameters are to be optimized
}

