% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/plantmethodsbasic.R
\docType{methods}
\name{phenology<-}
\alias{phenology<-}
\alias{phenology<-,Plant-method}
\title{Setting the phenology data.frame}
\usage{
phenology(object) <- value

\S4method{phenology}{Plant}(object) <- value
}
\arguments{
\item{object}{a Plant object}

\item{value}{a data.frame with the relevant phenology data. Given N stages,
the data.frame should have columns year, event1 through eventN+1, and
length1 through lengthN.}
}
\description{
Used to change the phenology data.frame without recreating the object.
}

