% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantmodelclass.R, R/plantmodelmethods.R
\docType{methods}
\name{phenology}
\alias{phenology}
\alias{phenology,PlantModel-method}
\title{Returns the phenology data.frame}
\usage{
phenology(object)

\S4method{phenology}{PlantModel}(object)
}
\arguments{
\item{object}{An object of class PlantModel}
}
\value{
A data.frame with the phenology data
}
\description{
Returns the phenology data.frame

Accesses the phenology data.frame of a PlantModel object
}

