% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelcheck.R
\name{phenologycheck}
\alias{phenologycheck}
\title{Checks phenology data.frame}
\usage{
phenologycheck(n, df, flower)
}
\arguments{
\item{n}{integer, number of stages in the model}

\item{df}{data.frame, the data.frame containing the phenology data.}

\item{flower}{logical, is the model a FlowerModel?}
}
\value{
This function returns TRUE if all the variables necessary to fit a
    PlantModel are present. If not it returns a vector with the logical FALSE
    and an error message with a list of the missing variables.
}
\description{
This function checks to see if the phenology data.frame has all of the
    correct columns.
}
