% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterfunctions.R
\name{pickEnding}
\alias{pickEnding}
\title{Picks correct ordinal indicator for a number}
\usage{
pickEnding(n)
}
\arguments{
\item{n}{numeric or integer, a number that needs an ordinal indicator.}
}
\value{
character, the ordinal indicator of the number provided.
}
\description{
This function returns the correct ordinal indicator for the number provided,
    i.e. the two letter suffix used to denote that a number is a rank
    (ex. 1st).
}
