% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Plant-class}
\alias{Plant-class}
\title{An S4 class to represent the development of a plant}
\description{
An S4 class to represent the development of a plant
}
\section{Slots}{

\describe{
\item{\code{phenology}}{A data.frame that contains the phenological data.}

\item{\code{temperature}}{A list that contains temperature data for all of
the years for which there is phenological data.}

\item{\code{modeltype}}{The type of structure the model will have. Options are
'partial', 'full', 'combined', or 'time'.}

\item{\code{form}}{The functional form of the thermal time accumulation.}

\item{\code{stages}}{An integer with the number of stages in the phenological model}
}}

