% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{plant}
\alias{plant}
\title{Creates Plant object}
\usage{
plant(pheno, clim, model, form, stages, parameters)
}
\arguments{
\item{pheno}{data.frame that contains the phenology data. Variables must
include year,}

\item{clim}{list or data.frame that contains the temperature data. Variables
must be year, day (julian), and temp (or tmin and tmax depending on the
 model).}

\item{model}{character, the name of the type of model to be used. Options are
partial, full. and combined. See Details for more information.}

\item{form}{function, the functional form of the thermal time calculations.
See details for more information}

\item{stages}{numeric, the number of phenological stages to be modeled.}

\item{parameters}{object of the class ParameterList. More information at
\code{\link{parameterlist}}.}
}
\description{
Creates Plant object
}

