% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantmodelclass.R
\docType{class}
\name{PlantModel-class}
\alias{PlantModel-class}
\title{An S4 class a fitted model of plant development}
\description{
An S4 class a fitted model of plant development
}
\section{Slots}{

\describe{
\item{\code{parameters}}{A list of ParameterList objects that contains the fit
parameters.}

\item{\code{error}}{A vector of RMSE values for each of the models (not cross-
validated).}

\item{\code{phenology}}{A data.frame that contains the phenological data.}

\item{\code{temperature}}{A list that contains temperature data for all of
the years for which there is phenological data.}

\item{\code{olm}}{A list of linear models relating the thermal time or day
accumulation to the stage length}

\item{\code{crossvalidated}}{logical, is the error crossvalidated?}
}}

