% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantmodel.R
\name{plantmodel}
\alias{plantmodel}
\title{Fits phenological models}
\usage{
plantmodel(phenology, temp, parlist, lbounds, ubounds, cores = 1L,
  estimateCT = TRUE, estimatelength = TRUE, simple = FALSE)
}
\arguments{
\item{phenology}{data.frame, contains the phenological data.
Required columns are `year`, and `event1` through `eventN` where `N`
is the number of phenological events in the model.}

\item{lbounds}{numeric, a vector of lower bounds for the parameters in the
model}

\item{ubounds}{numeric, a vector of the upper bounds for the parameters in
the model.}

\item{cores}{integer, if using parallel processing, how many cores should R
use to fit the model.}

\item{estimateCT}{logical, should the cardinal temperatures be estimated?}

\item{estimatelength}{logical, should the accumulation length or threshold be
estimated?}

\item{temps}{data.frame, contains the temperature data. There should be a
column for year, day of the year, hourly temperature (named temp) or min
max daily temperatures (named tmin and tmax), and hour if the
temperatures are hourly.}

\item{parameters}{ParameterList, contains the parameter values and
functional form of the thermal time calculations.}

\item{simplified}{logical, should the simplified version of the model be run.}
}
\value{
A PlantModel object.
}
\description{
`plantmodel` fits phenological models of two general types: thermal time accumulation and day accumulation (for more information see ____).
}

