% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantmodelcheck.R
\name{plantmodelcheck}
\alias{plantmodelcheck}
\title{Checks a number of problems with the PlantModel}
\usage{
plantmodelcheck(phenology, temperature, form, modtype, stages)
}
\arguments{
\item{phenology}{data.frame, the data.frame containing the phenology data.}

\item{temperature}{list, the list of temperatures used to calculate thermal time
accumulation.}

\item{form}{character, the functional form of the thermal time model.}

\item{modtype}{character, the type of model to be fit using plantmodel()}

\item{stages}{integer, number of phenological stages in the model.}
}
\description{
Checks a number of problems with the PlantModel
}

