% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{predictevent}
\alias{predictevent}
\title{Finds the day the threshold is met}
\usage{
predictevent(pars, temps, form, length)
}
\arguments{
\item{pars}{cardinal temperatures or NA for utah model}

\item{temps}{vector of temperatures}

\item{form}{the functional form of the thermal time or chill accumulation}

\item{length}{the threshold that the model runs until}
}
\value{
A vector of days when the threshold is reached
}
\description{
This function calculates the thermal time or chill accumulated and then
    identifies what day the thermal time or chill threshold is met.
}
