% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowerPlant.R
\name{responseVar}
\alias{responseVar}
\title{Returns response variable name}
\usage{
responseVar(Model, stage)
}
\arguments{
\item{Model}{character, the type of model used, either PlantModel or
FlowerModel}

\item{stage}{numeric, what stage is the is the model predicting for.}
}
\value{
The name of the response variable as a character.
}
\description{
This function takes in the type of model (PlantModel or FlowerModel) and
    the stage and based on that decides what the name of the response
    variable to be predicted in the linear model.
}
