% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rmsd}
\alias{rmsd}
\title{Calculates the Root Mean Squared Deviation/Error}
\usage{
rmsd(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{numeric, the first vector}

\item{y}{numeric, the second vector}

\item{na.rm}{logical, should NAs be removed from the vectors?}
}
\value{
This function returns the RMSD of the two vectors.
}
\description{
This function calculates the root mean squared error/deviation (RMSD/E)
    between two vectors of numbers.
}
\examples{
a <- rnorm(10)
b <- rnorm(10)
rmsd(a,b)
}
