% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rmse}
\alias{rmse}
\title{Calculates the RMSE}
\usage{
rmse(predicted, observed)
}
\arguments{
\item{predicted}{numeric, the vector of predicted values}

\item{observed}{numeric, the vector of observed values}
}
\value{
The root mean squared error of the two numeric vectors.
}
\description{
Calculates the RMSE
}
\examples{
a <- rnorm(10)
b <- rnorm(10)
rmse(a,b)
}
