% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rowDifference}
\alias{rowDifference}
\title{Returns rows in x that are missing from y}
\usage{
rowDifference(x, y)
}
\arguments{
\item{x}{A data.frame with all the data}

\item{y}{A data.frame that may have some rows missing}
}
\value{
the rows of x that are missing in y
}
\description{
Returns rows in x that are missing from y
}

