% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterfunctions.R
\name{showparlist}
\alias{showparlist}
\title{Legwork behind the ParameterList show method}
\usage{
showparlist(object)
}
\arguments{
\item{object}{Parameterlist object}
}
\value{
A data.frame with the columns type, form, length, lims, stage, Base,
    Optimal, and Critical
}
\description{
Legwork behind the ParameterList show method
}
