% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R
\docType{methods}
\name{simplified}
\alias{simplified}
\alias{simplified,ParameterList-method}
\title{Is the model simplified?}
\usage{
simplified(object)

\S4method{simplified}{ParameterList}(object)
}
\arguments{
\item{object}{An object of class ParameterList}
}
\value{
logical, is the simplified version of the model being fit? See ____
     for more information.
}
\description{
Is the model simplified?

Accesses whether the model in the ParameterList object is simplified
}
