% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{stagecrossval}
\alias{stagecrossval}
\title{Crossvalidating one stage}
\usage{
stagecrossval(pdat, modtype, k, stage, fun)
}
\arguments{
\item{pdat}{data.frame, the phenology data.frame.}

\item{modtype}{character, the type of model, either 'thermal' or 'day'.}

\item{k}{numeric, the number of folds in the crossvalidation.}

\item{stage}{numeric, the stage of the phenology model.}

\item{fun}{character, the name of the function used to evaluate the model.}
}
\value{
The mean rmse
}
\description{
Crossvalidating one stage
}

