% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantclass.R, R/plantmethodsbasic.R
\docType{methods}
\name{stages<-}
\alias{stages<-}
\alias{stages<-,Plant-method}
\title{Setting the number of stages}
\usage{
stages(object) <- value

\S4method{stages}{Plant}(object) <- value
}
\arguments{
\item{object}{A Plant object}

\item{value}{The number of phenological stages in the model.}
}
\description{
Used to change the number of stages in the phenological model without
     recreating the object.
}

