% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R,
%   R/plantmodelmethods.R, R/flowermodelmethods.R
\docType{methods}
\name{stages}
\alias{stages}
\alias{stages,ParameterList-method}
\alias{stages,PlantModel-method}
\alias{stages,FlowerModel-method}
\title{Returns the number of stages in the model}
\usage{
stages(object)

\S4method{stages}{ParameterList}(object)

\S4method{stages}{PlantModel}(object)

\S4method{stages}{FlowerModel}(object)
}
\arguments{
\item{object}{An object of class ParameterList}
}
\value{
The number of stages in the model
}
\description{
Returns the number of stages in the model

Accesses the number stages of a ParameterList object

Accesses the number of stages of a PlantModel object

Accesses the number of stages of a FlowerModel object
}
