% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R
\docType{methods}
\name{stagetype}
\alias{stagetype}
\alias{stagetype,ParameterList-method}
\title{Returns whether the model is a 'PlantModel' or a 'FlowerModel'}
\usage{
stagetype(object)

\S4method{stagetype}{ParameterList}(object)
}
\arguments{
\item{object}{An object of class ParameterList}
}
\value{
character, whether the model is a PlantModel or a FlowerModel
}
\description{
Returns whether the model is a 'PlantModel' or a 'FlowerModel'

Accesses whether model is for PlantModel or FlowerModel
}
