% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R
\docType{methods}
\name{start<-}
\alias{start<-}
\alias{start<-,ParameterList-method}
\title{Setting the start day of the model}
\usage{
start(object) <- value

\S4method{start}{ParameterList}(object) <- value
}
\arguments{
\item{object}{An object of class ParameterList}

\item{value}{A vector of start days for the model or NAs for starting at
bloom for PlantModels and harvest for FlowerModels}
}
\description{
Used to change the start day without recreating the object.
}
