% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R
\docType{methods}
\name{start}
\alias{start}
\alias{start,ParameterList-method}
\title{Returns the returns a vector of start days}
\usage{
start(object)

\S4method{start}{ParameterList}(object)
}
\arguments{
\item{object}{An object of class ParameterList}
}
\value{
A vector of the start days or NAs for starting at bloom for
    PlantModels and harvest for FlowerModels
}
\description{
Returns the returns a vector of start days

Accesses the start day of the ParameterList object
}
