% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantmodelcheck.R
\name{tempclasscheck}
\alias{tempclasscheck}
\title{Checks temperature class}
\usage{
tempclasscheck(frm, temp)
}
\arguments{
\item{frm}{character, the functional form of the thermal time model.}

\item{temp}{data.frame, the data.frame with temperatures used to calculate
thermal time accumulation.}
}
\value{
This function returns True if the class of the temperature is
    correct. Otherwise it returns a vector with FALSE and an error message.
}
\description{
This function checks to make sure the temperature data is in the right format
    given the functional form used to calculate thermal time with it.
}
