% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantclass.R, R/plantmethodsbasic.R
\docType{methods}
\name{temperature<-}
\alias{temperature<-}
\alias{temperature<-,Plant-method}
\title{Setting the temperature list}
\usage{
temperature(object) <- value

\S4method{temperature}{Plant}(object) <- value
}
\arguments{
\item{object}{A Plant object}

\item{value}{A list with the relevant temperature data. If the form
of the model is 'gdd' or 'gddsimple' the list should have the
variables tmin and tmax. Otherwise it should have vectors of hourly
temperatures.}
}
\description{
Used to change the temperature list without recreating the object.
}

