% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantmodelclass.R, R/plantmodelmethods.R
\docType{methods}
\name{temperature}
\alias{temperature}
\alias{temperature,PlantModel-method}
\title{Returns the temperature list}
\usage{
temperature(object)

\S4method{temperature}{PlantModel}(object)
}
\arguments{
\item{object}{An object of class PlantModel}
}
\value{
A list with the temperature data
}
\description{
Returns the temperature list

Accesses the temperature data.frame of a PlantModel object
}

