% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R
\docType{methods}
\name{threshold<-}
\alias{threshold<-}
\alias{threshold<-,ParameterList-method}
\title{Setting the model threshold}
\usage{
threshold(object) <- value

\S4method{threshold}{ParameterList}(object) <- value
}
\arguments{
\item{object}{An object of class ParameterList}

\item{value}{The time or thermal time threshold of the model.}
}
\description{
Used to change the model threshold without recreating the object.
}
