% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R,
%   R/plantmodelmethods.R, R/flowermodelmethods.R
\docType{methods}
\name{threshold}
\alias{threshold}
\alias{threshold,ParameterList-method}
\alias{threshold,PlantModel-method}
\alias{threshold,FlowerModel-method}
\title{Returns the returns a vector of thresholds}
\usage{
threshold(object)

\S4method{threshold}{ParameterList}(object)

\S4method{threshold}{PlantModel}(object)

\S4method{threshold}{FlowerModel}(object)
}
\arguments{
\item{object}{An object of class ParameterList}
}
\value{
The time or thermal time thresholds for the model.
}
\description{
Returns the returns a vector of thresholds

Accesses the threshold vector of a ParameterList object

Accesses the model lengths of a plant object

Accesses the model lengths of a FlowerModel object
}
