% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{timeSeriesCheck}
\alias{timeSeriesCheck}
\title{Checks time series for missing entries}
\usage{
timeSeriesCheck(x, start = NA, end = NA, hours = FALSE,
  datename = "dt", dtformat = "\%Y-\%m-\%d \%H:\%M:\%OS")
}
\arguments{
\item{x}{A data.frame that contains the time series. It must have at least
one column, named with the string in the datename argument.}

\item{start}{POSIXct or chr, the beginning date time.}

\item{end}{POSIXct or chr, the ending date time.}

\item{hours}{Logical, determines whether the function will check whether
all the hourly observations are there.}

\item{datename}{string, the name of the}

\item{dtformat}{The format the date time string is in.}
}
\value{
This function returns TRUE if there are no missing observations
    and returns a vector of the dates of the missing observations
    if there are any.
}
\description{
This function checks a time series, in the form of a data.frame, for days
    that are missing from the series.
}
