% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermal.R
\name{trapezoid}
\alias{trapezoid}
\title{Calculates thermal time based on the trapezoid model}
\usage{
trapezoid(Tvec, Tb, To, Ts, Tc, sum = TRUE)
}
\arguments{
\item{Tvec}{A vector of hourly temperatures, in degrees C.}

\item{Tb}{The base cardinal temperature of the model.}

\item{To}{The optimal cardinal temperature of the model.}

\item{Ts}{The subcritical cardinal temperature of the model.}

\item{Tc}{The critical cardinal temperature of the model.}

\item{sum}{Logical, should the vector of thermal times be returned as a
sum?}
}
\value{
either the number or vector of numbers representing the thermal
    time calculated
}
\description{
This function calculates thermal time in growing degree hours from a
    vector of temperatures (in degrees C) based on the trapezoid
    model of thermal time described below.
}
\details{
The functional form of the trapezoid thermal time model is as
    follows:
    \deqn{ GDH = \begin{cases}
        0 & T\leq T_b \\
        T - T_b & T_b \leq T \leq T_o \\
        T_o - T_b & T_o\leq T \leq T_s \\
        T*(Tb-To)/(Tc-Ts) + Tc*(To-Tb)/(To) & T_s\leq T \leq T_c \\
        0 & T_c \leq T
        \end{cases}}
    where \eqn{T} is the hourly temperature, \eqn{T_b} is the base
    temperature, \eqn{T_o} is the optimal temperature, and \eqn{T_c} is
    the critical temperature.
}
\examples{
temp <- seq(-5, 50)
gdh <- trapezoid(temp, 4, 25, 36, 40, sum=FALSE)
plot(gdh ~ temp)
}

