% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermal.R
\name{utahalt}
\alias{utahalt}
\title{Calculates winter chill based on the Utah Model}
\usage{
utahalt(tvec, sum = TRUE)
}
\arguments{
\item{tvec}{numeric, a vector of temperatures used to calculate chill}

\item{sum}{logical, should the accumulated chill be summed?}
}
\value{
A numeric vector of chill values or a single chill value if sum is
    TRUE.
}
\description{
This function calculates the amount of chill accumulated based on the Utah
    model.
}
\details{
This model is described in at \url{http://fruitsandnuts.ucdavis.edu/Weather_Services/chilling_accumulation_models/about_chilling_units/}
}
