% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R
\docType{methods}
\name{varyingpars<-}
\alias{varyingpars<-}
\alias{varyingpars<-,ParameterList-method}
\title{Setting the varying parameters}
\usage{
varyingpars(object) <- value

\S4method{varyingpars}{ParameterList}(object) <- value
}
\arguments{
\item{object}{An object of class ParameterList}

\item{value}{A vector of parameters that will vary from year to year}
}
\description{
Used to change the parameters that will vary without recreating the object.
}
