% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{whichtemp}
\alias{whichtemp}
\title{Returns the right temperature list}
\usage{
whichtemp(form, daily, hourly)
}
\arguments{
\item{form}{character, what is the form of the model}

\item{daily}{list the daily temperature data}

\item{hourly}{list, the hourly temperature data}
}
\value{
A list of temperature data based on the functional form of the model.
}
\description{
Returns the right temperature list
}

