% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwards.R
\name{yearconversion}
\alias{yearconversion}
\title{Converts years for bloom modeling}
\usage{
yearconversion(fdat, firstyear = NA, lastyear = NA, bloomvar = "event1",
  matvar = "event2", id.vars = NA)
}
\arguments{
\item{fdat}{dataframe, contains the phenology data}

\item{firstyear}{numeric, the first year you would like to predict bloom.}

\item{lastyear}{numeric, the last year you would like to predict bloom.}

\item{bloomvar}{character, the name of the column that contains the bloom
data.}

\item{matvar}{character, the name of the column that contains the harvest
data.}

\item{id.vars}{character, the name of columns that you would like to use to
identify the different observations (ex. ID or cultivar). These must be
these must be present in the fdat data.frame.}
}
\value{
A data.frame with all of the converted phenology data in it.
}
\description{
Models for bloom necessarily incorporte phenology data from multiple years.
    this is not compatible with the current phenoclim model paradigm. So this
    function groups the phenogical events by blooming event not by year.
}
