% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwards.R
\name{yearflip}
\alias{yearflip}
\title{Moves temperature data a year later}
\usage{
yearflip(year, tdat, start, hourly = TRUE)
}
\arguments{
\item{year}{numeric, the year you need to add temperature data to}

\item{tdat}{data.frame, the object that stores all your temperature data.}

\item{start}{numeric, the day of the year to go back to in the previous year.}

\item{hourly}{logical, is the data hourly?}
}
\value{
A data.frame that contains all of the temperature data points with
    negative days
}
\description{
This function takes temperature data and converts the year to the next year
    and also converts the days to be negative, counting back from 0 as
    Dec 31.
}
