% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipday.R
\name{yearfliptemp}
\alias{yearfliptemp}
\title{Moves temperature data a year later}
\usage{
yearfliptemp(year, tdat, start, hourly = TRUE)
}
\arguments{
\item{year}{numeric, the year you need to add temperature data to}

\item{tdat}{data.frame, the object that stores all your temperature data.}

\item{start}{numeric, the day of the year to go back to in the previous year.}

\item{hourly}{logical, is the data hourly?}
}
\value{
A data.frame that contains all of the temperature data points with
    negative days as well as day
}
\description{
This function takes temperature data and converts the year to the next year,
    converts the days to be negative (counting back from 0 as Dec 31), and
    creates a day index to use in extracting temps. The day index sets the
    start date to be 1.
}
