% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowerPlant.R
\name{yearlength}
\alias{yearlength}
\title{Calculates number of days in a year}
\usage{
yearlength(y)
}
\arguments{
\item{y}{numeric a vector of years}
}
\value{
A numeric vector of the lengths of the years inputted.
}
\description{
This function determines whether or not a given year is a leap year anc then
    if it is returns 366, and if not returns 365.
}
