% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearsums.R
\name{yearsums}
\alias{yearsums}
\title{Calculates thermal time sums by year}
\usage{
yearsums(pars, fdat, tdat, type, tempname = "temp", sumlen = NA,
  fn = "flower", hn = "harvest")
}
\arguments{
\item{pars}{lkasjsdf}

\item{fdat}{lkasjsdf}

\item{tdat}{lkasjsdf}

\item{type}{lkasjsdf}

\item{tempname}{lkasjsdf}

\item{sumlen}{lkasjsdf}

\item{fn}{lkasjsdf}

\item{hn}{lkasjsdf}
}
\value{
thermal sums
}
\description{
Calculates thermal time sums by year
}

