% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{avgdate}
\alias{avgdate}
\title{Averages a vector of dates}
\usage{
avgdate(dates, origin = "1970-01-01")
}
\arguments{
\item{dates}{a vector of dates}

\item{origin}{The start date. For R it is "1970-01-01".}
}
\value{
The average of all the dates
}
\description{
Averages a vector of dates
}
\examples{
v <- runif(10, min=9014, max=20300)
d <- as.Date(v, origin="1970-01-01")
avgdate(d)
}

