% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterfunctions.R
\name{checkpars}
\alias{checkpars}
\title{Checks to see if parameters are in the correct order}
\usage{
checkpars(pars, start, end, mclass, thresh = NA, form = NA)
}
\arguments{
\item{pars}{a vector of cardinal temperatures, should be organized least to
greatest}

\item{start}{numeric, the day the model starts running}

\item{end}{numeric, the day of the event to be predicted}

\item{mclass}{character, the model class, "FlowerModel" or "PlantModel"}

\item{thresh}{numeric, the threshold day for a DT model.}

\item{form}{character, the functional form used to calculate thermal time.}
}
\value{
logical, TRUE if the cardinal temperatures are organized least to greatest, and start and thresh comes before predicted event. FALSE otherwise
}
\description{
Checks to see if parameters are in the correct order
}
