% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{checktempyears}
\alias{checktempyears}
\title{Checks year presence in temperature list}
\usage{
checktempyears(pheno, temperature, modelclass)
}
\arguments{
\item{pheno}{data.frame, the data.frame containing the phenology data.}

\item{temperature, }{list, the list containing all the temperature data.}

\item{modelclass}{character, is the model a FlowerModel or a PlantModel}
}
\value{
Logical. Returns TRUE if all years of phenology data also have
    corresponding temperature data. Otherwise it returns the years that are
    missing temperature data.
}
\description{
This function checks to make sure that for every year where there is
    phenology data there is also temperature data.
}
