% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantmodelclass.R, R/plantmodelmethods.R,
%   R/flowermodelmethods.R
\docType{methods}
\name{crossval}
\alias{crossval}
\alias{crossval,PlantModel-method}
\alias{crossval,FlowerModel-method}
\title{Crossvalidates a Plant or Flower Model object}
\usage{
crossval(object, ...)

\S4method{crossval}{PlantModel}(object, temps, k, seed, fun, lbounds,
  ubounds, iterations = 100, cores = 1L)

\S4method{crossval}{FlowerModel}(object, temps, k, seed, fun, lbounds,
  ubounds, iterations = 100, cores = 1L)
}
\arguments{
\item{object}{an (non-crossvalidated) object of class PlantModel or
FlowerModel}

\item{...}{see other documentation}

\item{temps}{data.frame, contains all the temperature data.}

\item{k}{numeric, number of folds for crossvalidation.}

\item{seed}{numeric, the number to be used as the seed. Set to NA for no seed
set.}

\item{fun}{character, the name of the function used to evaluate the model.}

\item{lbounds}{numeric, lower bounds of parameters}

\item{ubounds}{numeric, upper bounds of parameters.}

\item{iterations}{numeric, number of iterations to be used when optimizing
paramters.}

\item{cores}{integer, number of cores to be used when fitting the models}
}
\value{
An object of class Plantmodel or FlowerModel with a crossvalidated
    error.
}
\description{
Crossvalidates a Plant or Flower Model object

Crossvalidate model

Crossvalidate model
}
