% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{crossval}
\alias{crossval}
\title{Crossvalidates the PlantModel Error}
\usage{
crossval(plant, temps, k, seed, fun = "rmsd")
}
\arguments{
\item{plant}{PlantModel, the object to be crossvalidated}

\item{temps}{data.frame, contains all the temperature data.}

\item{k}{numeric, number of folds for crossvalidation.}

\item{seed}{numeric, the number to be used as the seed. Set to NA for no seed
set.}

\item{fun}{character, the name of the function used to evaluate the model.}
}
\value{
A PlantModel object with the error slot having the new crossvalidated
    error
}
\description{
Does k-fold crossvalidation on the error of a PlantModel object.
}

