% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermalsum.R
\name{DTsum}
\alias{DTsum}
\title{Calculates thermal time sums for DT model}
\usage{
DTsum(pars, fdat, tdat, form, length, stage)
}
\arguments{
\item{pars}{Cardinal temperatures}

\item{fdat}{the data.frame containing the phenological information}

\item{tdat}{list containing the temperature information}

\item{form}{the functional form of the thermal time accumulation}

\item{length}{the length of thermal time accumulation (in days). It can be
either a set length of time (one number) or the total length of the
stage (one length for each entry in fdat).}

\item{stage}{the number of the stage of the phenological model}
}
\value{
The thermal sums for a given series of years.
}
\description{
Calculates thermal time sums for a given length of days after flowering for
a series of years.
}
