% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{eventi}
\alias{eventi}
\title{returns a column of the phenology data.frame}
\usage{
eventi(dat, i)
}
\arguments{
\item{dat}{the phenology data frame}

\item{i}{the number of the phenology event to be extracted}
}
\value{
A vector with the julian days of the ith phenological event
}
\description{
returns a column of the phenology data.frame
}

