% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractResults.R
\name{extractResults}
\alias{extractResults}
\title{Gets phenologial data and season length predictions}
\usage{
extractResults(pm, cultivars, form)
}
\arguments{
\item{pm}{PlantModel or FlowerModel object}

\item{cultivars}{character, the cultivars you want to extract results for.}

\item{form}{character, the functional form you want to extract results for.}
}
\value{
A data.frame with the results.
}
\description{
This will get results from a PlantModel or FlowerModel object
}
