% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{extracttemp}
\alias{extracttemp}
\title{Extracts temperatures from a data frame}
\usage{
extracttemp(tdat, years, starts, ends, tempname = NA, yearname = "year",
  dayname = "day")
}
\arguments{
\item{tdat}{a data frame containing the temperature data.}

\item{years}{a vector of years over which to extract temperature data.}

\item{starts}{a number or vector of numbers between 1 and 365 (inclusive)
specifying what days of the year to begin the temperature data extraction.}

\item{ends}{a number or vector of numbers between 1 and 365 (inclusive)
specifying what days of the year to end the temperature data extraction.}

\item{tempname}{name of the column that contains the temperature data. The
function will try and detect the name of the column if tempname is left
as NA.}

\item{yearname}{the name of the column that specifies the year.}

\item{dayname}{the name of the column that specifies the day of the year,
numeric.}
}
\value{
\code{extracttemp} returns a list the same length as \code{years},
    where each element of the list is a vector of all the temperatures in
    between \code{starts} and \code{ends} that year.
}
\description{
This function extracts temperature data from a data frame for specific days
    in a year over a given set of years. It is used with the yearsums and
    minrmse functions.
}
\details{
If tempname is left as NA the function will try to detect the
    correct column (either temp, tmin or tmax). However, it is safer to just
    specify the correct column name.
}
