% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{extracttemplist}
\alias{extracttemplist}
\title{Extracts the right templist based on the forms provided}
\usage{
extracttemplist(temps, forms)
}
\arguments{
\item{temps}{data.frame, where all the temperature data is stored.}

\item{forms}{list or character, the functional forms that will be used to
calculate the thermal time.}
}
\value{
A list of two the first contains the daily extracted temps if they
    are needed (if not it contains NA) and the second element contains the
    hourly extracted temps if they are needed (if not it contains NA).
}
\description{
Extracts the right templist based on the forms provided
}

