% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R, R/plantmethodsbasic.R
\docType{methods}
\name{form<-}
\alias{form<-}
\alias{form<-,ParameterList-method}
\alias{form<-,Plant-method}
\title{Setting the form}
\usage{
form(object) <- value

\S4method{form}{ParameterList}(object) <- value

\S4method{form}{Plant}(object) <- value

\S4method{form}{Plant}(object) <- value
}
\arguments{
\item{object}{An object of class ParameterList}

\item{value}{A list of cardinal temperatures}
}
\description{
Used to change the functional form without recreating the object.
}

