% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/plantmethodsbasic.R
\docType{methods}
\name{form}
\alias{form}
\alias{form,Plant-method}
\title{Returns the functional form of the model}
\usage{
form(object)

\S4method{form}{Plant}(object)
}
\arguments{
\item{object}{An object of class Plant}
}
\value{
Returns the functional form of the thermal time accumulation model.
}
\description{
Returns the functional form of the model

Accesses the thermal time functional form of a Plant object
}

