% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R,
%   R/plantmodelmethods.R
\docType{methods}
\name{form}
\alias{form}
\alias{form,ParameterList-method}
\alias{form,PlantModel-method}
\title{Returns the functional form}
\usage{
form(object)

\S4method{form}{ParameterList}(object)

\S4method{form}{PlantModel}(object)
}
\arguments{
\item{object}{An object of class ParameterList}
}
\value{
character, the name of the functional form of the model of thermal
    time accumulation.
}
\description{
Returns the functional form

Accesses the form of a ParameterList object

Accesses form of a PlantModel object
}
