% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headlist.R
\name{head.list}
\alias{head.list}
\title{Alternate implementation of head for lists}
\usage{
\method{head}{list}(x, n = 6L, ...)
}
\arguments{
\item{x}{a list}

\item{n}{the number of elements of the list to display.}

\item{...}{other arguments to pass to \code{head}.}
}
\description{
Alternate implementation of head for lists
}

