% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R
\docType{methods}
\name{mclass<-}
\alias{mclass<-}
\alias{mclass<-,ParameterList-method}
\alias{mclass<-,list-method}
\title{Setting stages as PlantModels or a FlowerModels}
\usage{
mclass(object) <- value

\S4method{mclass}{ParameterList}(object) <- value

\S4method{mclass}{list}(object) <- value
}
\arguments{
\item{object}{An object of class ParameterList}

\item{value}{character, is the model to be fit a 'PlantModel' or a
'FlowerModel'.}
}
\description{
Used to change whether the model uses the PlantModel class or the
 FlowerModel class.
}
