% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/plantmodelmethods.R
\docType{methods}
\name{mclass}
\alias{mclass}
\alias{mclass,PlantModel-method}
\title{Returns whether the model is a 'PlantModel' or a 'FlowerModel'}
\usage{
mclass(object)

\S4method{mclass}{PlantModel}(object)
}
\arguments{
\item{object}{An object of class ParameterList}
}
\value{
character, whether the model is a PlantModel or a FlowerModel
}
\description{
Returns whether the model is a 'PlantModel' or a 'FlowerModel'

Accesses the model class of a PlantModel object
}
