% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minrmse.R
\name{minrmse}
\alias{minrmse}
\title{Calculates model RMSE}
\usage{
minrmse(pars, fdat, tdat, modtype, form, stage, CT, S, TH, simple, varying,
  modclass, firstevent)
}
\arguments{
\item{pars}{vector with threshold value and cardinal temperatures}

\item{fdat}{the data.frame containing the phenological information}

\item{tdat}{list containing the temperature information}

\item{modtype}{character, the type of model. The model can be 'DT' (day
threshold) or 'TTT' (thermal time threshold)}

\item{form}{the functional form of the thermal time accumulation}

\item{stage}{the number of the stage of the phenological model}

\item{CT}{logical, should the cardinal temperatures be optimized. If not CT
is the vector of cardinal temperatures.}

\item{S}{logical/numeric, is the day to start counting being optimized? If
not, the value associated with the start day should be provided.}

\item{TH}{logical/numeric, should the model threshold be optimized. If
not L is the (numeric) model threshold.}

\item{simple}{logical, should the simplified version of the model be run?}

\item{varying}{character, c('start', 'threshold') should either of these pars
vary from year to year.}

\item{modclass}{character, type of model to be estimating, options are
'PlantModel' or 'FlowerModel'. If you have negative day values, you
probably want flower model.}

\item{firstevent}{character, the name of the column that contains the
starting phenological event for the model. For FlowerModels this is the
harvest date.}
}
\value{
The RMSE value for a given set of cardinal temperatures start day
    model threshold.
}
\description{
Calculates RMSE for a Day Threshold (DT) or Thermal Time Threshold (TTT)
model of thermal time accumulation given a certain set of cardinal
temperatures and a day/thermal time threshold.
}
