% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minrmse.R
\name{minrmse}
\alias{minrmse}
\title{Calculates model RMSE}
\usage{
minrmse(pars, fdat, tdat, modtype, form, stage, CT, TH, simple, startday,
  modclass)
}
\arguments{
\item{pars}{vector with threshold value and cardinal temperatures}

\item{fdat}{the data.frame containing the phenological information}

\item{tdat}{list containing the temperature information}

\item{modtype}{character, the type of model. The model can be 'DT' (day
threshold) or 'TTT' (thermal time threshold)}

\item{form}{the functional form of the thermal time accumulation}

\item{stage}{the number of the stage of the phenological model}

\item{CT}{logical, should the cardinal temperatures be optimized. If not CT
is the vector of cardinal temperatures}

\item{TH}{logical, should the model accumulation threshold be optimized. If
not L is the (numeric) model accumulation threshold.}

\item{simple}{logical, should the simplified version of the model be run?}

\item{startday}{logical, is the day to start counting being optimized? (ie.
not using bloom as startday)}

\item{modclass}{logical, Do count forward from the starting event (as
opposed to backward)? If you have negative values in your event days
you forward should probably be FALSE. Bloom model forward=false, harvest
model forward=true.}
}
\value{
The RMSE value for a given set of cardinal temperatures and thermal
    time accumulation length.
}
\description{
Calculates RMSE for a Day Threshold (DT) or Thermal Time Threshold (TTT)
model of thermal time accumulation given a certain set of cardinal
temperatures and a day/thermal time threshold.
}
