% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minrmse.R
\name{minrmse}
\alias{minrmse}
\title{Calculates model RMSE}
\usage{
minrmse(pars, fdat, tdat, modtype, form, stage, CT, L, simple)
}
\arguments{
\item{pars}{vector with threshold value and cardinal temperatures}

\item{fdat}{the data.frame containing the phenological information}

\item{tdat}{list containing the temperature information}

\item{modtype}{character, the type of model. The model can be 'DT' (day
threshold) or 'TTT' (thermal time threshold)}

\item{form}{the functional form of the thermal time accumulation}

\item{stage}{the number of the stage of the phenological model}

\item{CT}{logical, should the cardinal temperatures be optimized. If not CT
is the vector of cardinal temperatures}

\item{L}{logical, should the model accumulation length be optimized. If not L
is the (numeric) model accumulation length.}

\item{simple}{logical, should the simplified version of the model be run?}
}
\value{
The RMSE value for a given set of cardinal temperatures and thermal
    time accumulation length.
}
\description{
Calculates RMSE for a Day Threshold (DT) or Thermal Time Threshold (TTT)
model of thermal time accumulation given a certain set of cardinal
temperatures and a day/thermal time threshold.
}

