% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minrmse.R
\name{minrmseTTTsimplified}
\alias{minrmseTTTsimplified}
\title{Calculates simplified TTT model RMSE}
\usage{
minrmseTTTsimplified(pars, fdat, tdat, form, thresh, stage, modclass)
}
\arguments{
\item{pars}{Cardinal temperatures}

\item{fdat}{the data.frame containing the phenological information}

\item{tdat}{list containing the temperature information}

\item{form}{the functional form of the thermal time accumulation}

\item{thresh}{the threshold in thermal time that days are counted for.}

\item{stage}{the number of the stage of the phenological model}

\item{modclass}{character, type of model to be estimating, options are
'PlantModel' or 'FlowerModel'. If you have negative day values, you
probably want flower model.}
}
\value{
The RMSE value for a given set of cardinal temperatures and thermal
    time accumulation length.
}
\description{
Calculates RMSE for the simplified model of day accumulation given a
    certain set of cardinal temperatures and day accumulation
    length. Counts the amount of thermal time required to reach the season
    length.
}
