% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R
\docType{methods}
\name{modeltype<-}
\alias{modeltype<-}
\alias{modeltype<-,ParameterList-method}
\title{Setting the model type}
\usage{
modeltype(object) <- value

\S4method{modeltype}{ParameterList}(object) <- value
}
\arguments{
\item{object}{An object of class ParameterList}

\item{value}{character, the model type, either 'thermal' or 'day'. See _____
for more information.}
}
\description{
Used to change the model type without recreating the object.
}
