% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantclass.R, R/plantmethodsbasic.R
\docType{methods}
\name{modeltype}
\alias{modeltype}
\alias{modeltype,Plant-method}
\title{Displays the model type}
\usage{
modeltype(object)

\S4method{modeltype}{Plant}(object)
}
\arguments{
\item{object}{An object of class Plant}
}
\value{
Returns the model type (character)
}
\description{
Displays the model type

Accesses the model type of a Plant object
}

