% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R,
%   R/plantmodelmethods.R
\docType{methods}
\name{modeltype}
\alias{modeltype}
\alias{modeltype,ParameterList-method}
\alias{modeltype,PlantModel-method}
\title{Returns the model type}
\usage{
modeltype(object)

\S4method{modeltype}{ParameterList}(object)

\S4method{modeltype}{PlantModel}(object)
}
\arguments{
\item{object}{An object of class ParameterList}
}
\value{
character, specifies what type of model the parameters are for.
    Options are 'thermal' and 'day'
}
\description{
Returns the model type

Accesses the model type of a ParameterList object

Accesses the model type of a PlantModel object
}
