% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R
\docType{methods}
\name{modlength}
\alias{modlength}
\alias{modlength,ParameterList-method}
\title{Returns the returns a vector of accumulation lengths}
\usage{
modlength(object)

\S4method{modlength}{ParameterList}(object)
}
\arguments{
\item{object}{An object of class ParameterList}
}
\value{
A vector of model lengths
}
\description{
Returns the returns a vector of accumulation lengths

Accesses the modlength vector of a ParameterList object
}

