% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectivefunction.R
\name{objective}
\alias{objective}
\title{Generates the Objective function}
\usage{
objective(parameters, phenology, temperature, stage, CT, L)
}
\arguments{
\item{parameters}{ParameterList, contains all of the parameters needed to
create the phenology model/PlantModel object.}

\item{phenology}{data.frame, contains all of the phenology data necessary for
running the phenology model.}

\item{temperature}{list, contains all the temperature data for estimating
thermal time.}

\item{stage}{numeric, the stage of the model.}

\item{CT}{Should the cardinal temperatures be optimized? If yes CT is TRUE,
if not CT should be a vector of the cardinal temperatures.}

\item{L}{Should accumulation length be optimized? If yes L is TRUE,
if not L should be the accumulation length to be used.}
}
\description{
This function creates a function to be passed to the DEoptim function that
optimizes the cardinal temperatures and the length of thermal time or day
accumulation for a PlantModel object.
}

