% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectivefunction.R
\name{objective}
\alias{objective}
\title{Generates the Objective function}
\usage{
objective(parlist, phenology, temp, stage, CT, Start, Threshold, listindex,
  mclass)
}
\arguments{
\item{parlist}{ParameterList, contains all of the parlist needed to
create the phenology model.}

\item{phenology}{data.frame, contains all of the phenology data necessary for
running the phenology model.}

\item{temp}{data.frame, contains all the temperature data for estimating
thermal time and has a POSIXct date-time column for indexing}

\item{stage}{numeric, the stage of the model.}

\item{CT}{Should the cardinal temperatures be optimized?}

\item{Start}{logical, Should model start day be optimized? If yes,
Start is TRUE.}

\item{Threshold}{logical, Should model threshold be optimized? If yes, it is
TRUE}

\item{listindex}{numeric, the index of the ParameterList from parlist that
you are at}

\item{mclass}{character, type of model to be estimating, options are
'PlantModel' or 'FlowerModel'. If you have negative day values, you
probably want flower model.}
}
\value{
the function that is passed to DEoptim to optimize.
}
\description{
This function creates a function to be passed to the DEoptim function that
optimizes the cardinal temperatures and the length of thermal time or day
accumulation for a PlantModel or FlowerModel object.
}
