% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectivefunction.R
\name{objective}
\alias{objective}
\title{Generates the Objective function}
\usage{
objective(parlist, phenology, templist, stage, CT, L, simple, listindex,
  startday, stgtype)
}
\arguments{
\item{parlist}{ParameterList, contains all of the parlist needed to
create the phenology model/PlantModel object.}

\item{phenology}{data.frame, contains all of the phenology data necessary for
running the phenology model.}

\item{templist}{list, contains all the temperature data for estimating
thermal time.}

\item{stage}{numeric, the stage of the model.}

\item{CT}{Should the cardinal temperatures be optimized? If yes CT is TRUE,
if not CT should be a vector of the cardinal temperatures.}

\item{L}{Should accumulation length be optimized? If yes L is TRUE,
if not L should be the accumulation length to be used.}

\item{simple}{logical, is the simplified version of the model being run?}

\item{listindex}{numeric, the index of the ParameterList from parlist that
you are at}

\item{startday}{logical, is the day to start counting being optimized? (ie.
not using bloom as startday)}

\item{stgtyp}{character, type of model to be estimating, options are
'PlantModel' or 'FlowerModel'. If you have negative day values, you
probably want flower model.}
}
\value{
the function that is passed to DEoptim to optimize.
}
\description{
This function creates a function to be passed to the DEoptim function that
optimizes the cardinal temperatures and the length of thermal time or day
accumulation for a PlantModel object.
}
