% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R
\docType{class}
\name{ParameterList-class}
\alias{ParameterList-class}
\title{An S4 class to store model parameters}
\description{
ParameterList stores parameters for running the phenological models
}
\section{Slots}{

\describe{
\item{\code{stages}}{numeric, The number of stages in the phenological model.}

\item{\code{modeltype}}{character. The type of phenological model to be used. Options
are 'DT' for Day threshold and 'TTT' for thermal time threshold. See ____
for more information.}

\item{\code{form}}{character. The functional form of the thermal time accumulation
model. Current options are gdd, gddsimple, linear, flat, triangle, and
anderson.}

\item{\code{cardinaltemps}}{list. The cardinal temperatures for the model. This is a
list that should either be of length one or the length of the number of
stages. Each element of the list should contain the same number of
cardinal parameters.}

\item{\code{threshold}}{vector. Stores the threshold of time or thermal time that is
accumulated in the model. Set to NA to run the base (simplified) model.}

\item{\code{start}}{numeric, stores the start day for accumulating time or thermal
time. If NA, model starts at event1 (bloom) for PlantModel and event0
(harvest) for FlowerModel.}

\item{\code{parsOptimized}}{character. Determines what parameters are optimized in
the model. `parsOptimized` is a character vector that can contain
"cardinaltemps", "threshold", "start" but it must contain at
least one (or else why optimize). Currently this cannot vary by stage.}

\item{\code{mclass}}{character. Is this a parameter list going to be used to
fit a 'PlantModel' or a 'FlowerModel'.}
}}

