% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R
\docType{class}
\name{ParameterList-class}
\alias{ParameterList-class}
\title{An S4 class to store model parameters}
\description{
ParameterList stores parameters for running the phenological models
}
\section{Slots}{

\describe{
\item{\code{stages}}{numeric, The number of stages in the phenological model.}

\item{\code{modeltype}}{character. The type of phenological model to be used. Options
are 'DT' for Day threshold and 'TTT' for thermal time threshold. See ____
for more information.}

\item{\code{form}}{character. The functional form of the thermal time accumulation
model. Current options are gdd, gddsimple, linear, flat, triangle,
anderson, basicchill, utah or utahalt.}

\item{\code{simplified}}{logical. Is the model to be fit a simplified model?}

\item{\code{cardinaltemps}}{list. The cardinal temperatures for the model. This is a
list that should either be of length one or the length of the number of
stages. Each element of the list should contain the same number of
cardinal parameters.}

\item{\code{threshold}}{numeric. Stores the threshold of time or thermal time that is
accumulated in the model. Set to NA to run the base (simplified) model.}

\item{\code{startday}}{numeric, stores the day of the year the model will start or
the offset to the previous stage end that the model starts. A startday
of 0 means either that the model starts the day the last stage ended.}

\item{\code{varyingpars}}{character, c('threshold', 'start'). the parameters that
will vary from year to year. If none will vary, set to NA.}

\item{\code{parsOptimized}}{character. Determines what parameters are optimized in
the model. `parsOptimized` is a character vector that can contain
"cardinaltemps", "threshold", "start" but it must contain at
least one (or else why optimize). Currently this cannot vary by stage.}

\item{\code{mclass}}{character. Is this a parameter list going to be used to
fit a 'PlantModel' or a 'FlowerModel'.}
}}

