% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{parameterlist}
\alias{parameterlist}
\title{Creates ParameterList object}
\usage{
parameterlist(n, mt, simple, ff, ct, modelthreshold, lims,
  optimized = c("cardinaltemps", "modlength"), ModelClass = "PlantModel")
}
\arguments{
\item{n}{integer, the number of stages in the model.}

\item{mt}{character, the model type. Can be either 'DT' or 'TTT'.}

\item{simple}{logical, is the model the simplified version?}

\item{ff}{character, the functional form of the thermal time calculations.
Options are 'gdd', 'gddsimple', 'linear', 'flat', 'triangle', and
'anderson'. For more information see ______.}

\item{ct}{A list, data.frame, or matrix of cardinal temperatures. In the
case of a data.frame or matrix, each row should contain the cardinal
temperatures for a given stage of the model. If ct is a list, each
element of the list should be a vector of cardinal temperatures for a
given stage of the model.}

\item{modelthreshold}{numeric vector of model thresholds.}

\item{lims}{list of start and stop pairs. Start is always the day close to
flowering. This means that it will be the earliest day in a PlantModel,
but the latest day in a FlowerModel. If length is not NA, one of the
limits has to be NA.}

\item{optimized}{character, Determines what parameters are optimized in the
model.Can contain "cardinaltemps", "modlength" or both, but it must
contain at least one of the two.}

\item{ModelClass}{character, the class to be fit. Options are
'PlantModel' or 'FlowerModel'.}
}
\value{
An object of the class ParameterList.
}
\description{
This function creates an object of the class ParameterList. Plant objects
    require objects of the class ParameterList for their parameters slot.
}
