% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/plantmethodsbasic.R
\docType{methods}
\name{parameters}
\alias{parameters}
\alias{parameters,Plant-method}
\title{Displays the model's initial parameter values}
\usage{
parameters(object)

\S4method{parameters}{Plant}(object)
}
\arguments{
\item{object}{An object of class Plant}
}
\value{
.
}
\description{
Displays the model's initial parameter values

Accesses the number of stages of a Plant object
}

