% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantmodelclass.R, R/plantmodelmethods.R
\docType{methods}
\name{parameters}
\alias{parameters}
\alias{parameters,PlantModel-method}
\title{Returns the parameters for a phenological model}
\usage{
parameters(object)

\S4method{parameters}{PlantModel}(object)
}
\arguments{
\item{object}{An object of class PlantModel}
}
\value{
The ParameterList of the phenological model
}
\description{
Returns the parameters for a phenological model

Accesses the parameters of a PlantModel object
}

