% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoclim.R
\docType{package}
\name{phenoclim}
\alias{phenoclim}
\alias{phenoclim-package}
\title{phenoclim: a package for phenological analysis and prediction}
\description{
phenoclim: a package for phenological analysis and prediction
}
