% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantclass.R, R/plantmethodsbasic.R
\docType{methods}
\name{phenology}
\alias{phenology}
\alias{phenology,Plant-method}
\title{Returns the phenology data.frame}
\usage{
phenology(object)

\S4method{phenology}{Plant}(object)
}
\arguments{
\item{object}{An object of class Plant}
}
\value{
A data.frame with the phenology data
}
\description{
Returns the phenology data.frame

Accesses the phenology data.frame of a Plant object
}

