% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantmodel.R
\name{plantmodel}
\alias{plantmodel}
\title{Fits phenological models}
\usage{
plantmodel(phenology, temps, parlist, lbounds, ubounds, cores = 1L,
  iterations = 200, small = FALSE)
}
\arguments{
\item{phenology}{data.frame, contains the phenological data.
Required columns are `year`, and `event1` through `eventN` where `N`
is the number of phenological events in the model.}

\item{temps}{data.frame, contains the temperature data. There should be a
column for year, day of the year, hourly temperature (named temp) or min
max daily temperatures (named tmin and tmax), and hour if the
temperatures are hourly.}

\item{parlist}{list of ParameterLists, contains the parameter values and
functional form of the thermal time calculations. Note as of right now.
All ParameterLists must have the same number of stages, and the same
model type.}

\item{lbounds}{numeric, a vector of lower bounds for the parameters in the
model}

\item{ubounds}{numeric, a vector of the upper bounds for the parameters in
the model.}

\item{cores}{integer, if using parallel processing, how many cores should R
use to fit the model.}

\item{iterations}{numeric, the number of iterations used in the differential
evolution optimization of the phenological parameters.}

\item{small}{logical, if small empty dataframes will be returned.}
}
\value{
A PlantModel object.
}
\description{
`plantmodel` fits phenological models of two general types: thermal time accumulation and day accumulation (for more information see ____).
}

