% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantmodel.R
\name{plantmodel}
\alias{plantmodel}
\title{Fits phenological models}
\usage{
plantmodel(phenology, temperature, model, parameters, lbounds, ubounds,
  cores = 1L, estimateCT = TRUE, estimatelength = TRUE,
  simplified = FALSE)
}
\arguments{
\item{phenology}{data.frame, contains the phenological data.
Required columns are `year`, and `event1` through `eventN` where `N`
is the number of phenological events in the model.}

\item{temperature}{list, contains the temperature data. Each element of
the list should be either a vector of hourly temperatures or a
data.frame of minimum and maximum daily temperatures. Each element
of the list should be named with the corresponding year.}

\item{model}{character, specifies the type of model. Options are 'tta'
(thermal time accumulation) or 'da' (day accumulation).}

\item{parameters}{ParameterList, contains the parameter values and
functional form of the thermal time calculations.}

\item{lbounds}{numeric, a vector of lower bounds for the parameters in the
model}

\item{ubounds}{numeric, a vector of the upper bounds for the parameters in
the model.}

\item{cores}{integer, if using parallel processing, how many cores should R
use to fit the model.}

\item{estimateCT}{logical, should the cardinal temperatures be estimated?}

\item{estimatelength}{logical, should the accumulation length or threshold be
estimated?}

\item{simplified}{logical, should the simplified version of the model be run.}
}
\description{
`plantmodel` fits phenological models of two general types: thermal time accumulation and day accumulation (for more information see ____).
}

