% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R,
%   R/plantmodelmethods.R, R/flowermodelmethods.R
\docType{methods}
\name{simplified}
\alias{simplified}
\alias{simplified,ParameterList-method}
\alias{simplified,list-method}
\alias{simplified,PlantModel-method}
\alias{simplified,FlowerModel-method}
\title{Returns whether the model is simplified}
\usage{
simplified(object)

\S4method{simplified}{ParameterList}(object)

\S4method{simplified}{list}(object)

\S4method{simplified}{PlantModel}(object)

\S4method{simplified}{FlowerModel}(object)
}
\arguments{
\item{object}{An object of class ParameterList}
}
\value{
logical, return whether the model to be fit is a simplified model
}
\description{
Returns whether the model is simplified

Accesses the simplified parameter of a ParameterList object

Accesses the simplified parameter of a list of ParameterLists

Accesses simplified parameter of a PlantModel object

Accesses the simplified parameter of a FlowerModel object
}
