% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R
\docType{methods}
\name{stages<-}
\alias{stages<-}
\alias{stages<-,ParameterList-method}
\title{Setting the number of stages}
\usage{
stages(object) <- value

\S4method{stages}{ParameterList}(object) <- value
}
\arguments{
\item{object}{An object of class ParameterList}

\item{value}{numeric, the number of stages in teh new model}
}
\description{
Used to change the number of stages without recreating the object.
}
