% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/plantmethodsbasic.R
\docType{methods}
\name{stages}
\alias{stages}
\alias{stages,Plant-method}
\title{Displays the number of stages in the phenological model}
\usage{
stages(object)

\S4method{stages}{Plant}(object)
}
\arguments{
\item{object}{An object of class Plant}
}
\value{
The number of stages in the phenological model (numeric)
}
\description{
Displays the number of stages in the phenological model

Accesses the number of stages of a Plant object
}

