% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistmethods.R
\docType{methods}
\name{mclass,ParameterList-method}
\alias{mclass,ParameterList-method}
\title{Accesses whether model is for PlantModel or FlowerModel}
\usage{
\S4method{mclass}{ParameterList}(object)
}
\description{
Accesses whether model is for PlantModel or FlowerModel
}
