% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipday.R
\name{startEnd}
\alias{startEnd}
\title{Creates day or hour counts}
\usage{
startEnd(start, stglength, hourly = TRUE, stgtype)
}
\arguments{
\item{start}{num, vector of starting days.}

\item{stglength}{num, a stage length or vector of stage lengths}

\item{hourly}{logical, is the model an hourly model}

\item{stgtype}{character, is the model a PlantModel or a FlowerModel?
Options are 'PlantModel' or 'FlowerModel'.}

\item{startday}{logical, is the start day being estimated in this model, as
opposed to using the bloom date.}
}
\value{
a vector of indexes that can be used to extract temperature data.
}
\description{
This function creates day and hour indexes to ube used to extract temperature
    data. It has options to work for harvest and bloom models.
}
