% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/plantmethodsbasic.R
\docType{methods}
\name{temperature}
\alias{temperature}
\alias{temperature,Plant-method}
\title{Returns the temperature list}
\usage{
temperature(object)

\S4method{temperature}{Plant}(object)
}
\arguments{
\item{object}{An object of class Plant}
}
\value{
A list with the temperature data
}
\description{
Returns the temperature list

Accesses the temperature data.frame of a Plant object
}

