% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipday.R
\name{tempyearconversion}
\alias{tempyearconversion}
\title{Creates temp data for FlowerModel}
\usage{
tempyearconversion(tdat, start, hourly = TRUE)
}
\arguments{
\item{tdat}{data.frame, the object that stores all your temperature data.}

\item{start}{numeric, the day of the year to go back to in the previous year.}

\item{hourly}{logical, is the data hourly?}
}
\value{
A data.frame that contains all of the temperature data points with
    negative days and a day index that starts from 1.
}
\description{
This function takes temperature data and reformats it so that it is
    compatible with the FlowerModel model structure. This means
    extracting temperatures from the previous year and adding them to the
    data but with negative numbers for days and a day index.
}
