% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermalsum.R
\name{thermalsum}
\alias{thermalsum}
\title{Calculates thermal time sums}
\usage{
thermalsum(pars, fdat, tdat, modtype, form, length, stage, forward = TRUE)
}
\arguments{
\item{pars}{Cardinal temperatures}

\item{fdat}{the data.frame containing the phenological information}

\item{tdat}{list containing the temperature information}

\item{modtype}{character, specifies what type of model is being run. Can be
either DT (Day Threshold) or TTT (Thermal Time Threshold).}

\item{form}{the functional form of the thermal time accumulation}

\item{length}{the length of thermal time accumulation (in days). It can be
either a set length of time (one number) or the total length of the
stage (one length for each entry in fdat).}

\item{stage}{the number of the stage of the phenological model}

\item{forward}{logical, Do count forward from the starting event (as
opposed to backward)? If you have negative values in your event days
you forward should probably be FALSE.}
}
\value{
The thermal sums for a given series of years.
}
\description{
Calculates thermal time sums for either the TTT or DT time model
}
