% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermalsum.R
\name{thermalsum}
\alias{thermalsum}
\title{Calculates thermal time sums}
\usage{
thermalsum(ctemps, yrs, tdat, modtype, form, start, thresh, varying, mclass)
}
\arguments{
\item{ctemps}{Cardinal temperatures}

\item{yrs}{numeric, the years we have phenology data for.}

\item{tdat}{list containing the temperature information}

\item{modtype}{character, specifies what type of model is being run. Can be
either DT (Day Threshold) or TTT (Thermal Time Threshold).}

\item{form}{the functional form of the thermal time accumulation}

\item{start}{POSIXct, the date to start accumulating time or thermal
time towards the model threshold.}

\item{thresh}{the length of thermal time accumulation (in either
days or thermal time units).}

\item{varying}{character, c('start', 'threshold') should either of these
pars vary from year to year.}

\item{mclass}{character, type of model to be estimating, options are
'PlantModel' or 'FlowerModel'.}
}
\value{
The thermal sums for a given series of years.
}
\description{
Calculates thermal time sums for either the TTT or DT time model
}
