% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{timeSeriesCheck}
\alias{timeSeriesCheck}
\title{Checks time series for missing entries}
\usage{
timeSeriesCheck(x, years = NA, limits = NA, hours = FALSE)
}
\arguments{
\item{x}{A data.frame that contains the time series. It must have at least
two columns, one named 'year' with the year, and one named 'day' with
the julian day. If hours is TRUE, it needs a column}

\item{years}{A numeric vector of years to check. If NA the function will
check all the years in the data.frame.}

\item{limits}{A numeric vector of length two giving the minimum and
maximum julian day to check. If left NA, the function will check all
the days of the year.}

\item{hours, }{Logical, determines whether the function will check whether
all the hourly observations are there}
}
\value{
This function returns TRUE if there are no missing observations
    and returns a list of the days (and hours) of the missing observations
    if there are any.
}
\description{
This function checks a time series, in the form of a data.frame, for days
    that are missing from the series.
}

