% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterlistclass.R, R/parameterlistmethods.R
\docType{methods}
\name{varyingpars}
\alias{varyingpars}
\alias{varyingpars,ParameterList-method}
\alias{varyingpars,list-method}
\title{Returns the parameters that vary from year to year}
\usage{
varyingpars(object)

\S4method{varyingpars}{ParameterList}(object)

\S4method{varyingpars}{list}(object)
}
\arguments{
\item{object}{An object of class ParameterList}
}
\value{
A vector of the parameters that will vary from year to year for
    PlantModels and harvest for FlowerModels
}
\description{
Returns the parameters that vary from year to year

Accesses the varyingpars vector of a ParameterList object

Accesses the varyingpars vector for a list of ParameterLists
}
